requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Google Gemini AI
google-generativeai==0.3.2

# Social Media APIs
tweepy==4.14.0
praw==7.7.1

# News APIs
newsapi-python==0.2.7
requests==2.31.0
aiohttp==3.9.1

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utilities
numpy==1.26.3
pandas==2.1.4
asyncio==3.4.3
cachetools==5.3.2
tenacity==8.2.3

# Logging and monitoring
structlog==24.1.0
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
